Stopped using 'long double' in jacobian calculator and tracer to avoid ambiguous overloads on windows.

+++ myokit/_sim/jacobian.cpp
@@ -57,7 +57,7 @@ tab = '    '
 #define N_INPUT <?= len(inputs) ?>
 
 // Define numerical type
-typedef long double Real;
+typedef double Real;
 
 // Number of derivatives in each derivative vector
 #define N_DIFFS <?= model.count_states() ?>
